Former-commit-id: dbf3c85a4f7c20a7716d2bcb6a73b4ba283ffe71

+++ source/gui/navigationWidgets/cxPlaybackWidget.h
@@ -70,7 +70,7 @@ private:
 	std::vector<TimelineEvent> convertRegistrationHistoryToEvents(ssc::RegistrationHistoryPtr reg);
 	void showDetails();
 	QColor generateRandomToolColor() const;
-	QString convertMillisecsToNiceString(int ms) const;
+	QString convertMillisecsToNiceString(qint64 ms) const;
 
 	QLabel* mLabel;
 	QLabel* mTotalLengthLabel;
